+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ###Version: 0.11.7
 - ```new``` Landscape layout for calculator. (**@i906**)
+- ```fix``` Can edit main currency.
 
 ###Version: 0.11.6
 - ```new``` Tags are ordered alphabetically.